修复

+++ Sources/JXPagingView/JXPagingSmoothView.swift
@@ -61,7 +61,7 @@ open class JXPagingSmoothView: UIView {
     var currentListInitializeContentOffsetY: CGFloat = 0
     var singleScrollView: UIScrollView?
     /// 顶部固定sectionHeader的垂直偏移量。数值越大越往下沉。
-    var pinSectionHeaderVerticalOffset: Int = 0
+    var pinSectionHeaderVerticalOffset: CGFloat = 0
 
     deinit {
         listDict.values.forEach {
